# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Help Desk API
  description: |
    API для системи управління заявками Help Desk
    
    ## Автентифікація
    API використовує JWT токени для автентифікації. Отримайте токен через `/api/auth/login` та передавайте його в заголовку `Authorization: Bearer <token>`.
    
    ## Обмеження швидкості
    - Загальні запити: 100 запитів на хвилину
    - Автентифікація: 5 спроб на хвилину
    - Завантаження файлів: 10 запитів на хвилину
    
  version: 1.0.0
  contact:
    name: Help Desk Support
    email: support@helpdesk.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.helpdesk.com/api
    description: Production server
...
paths:
...
  /tickets/{id}/comments:
    get:
      tags:
        - Comments
      summary: Отримання коментарів тикету
      description: Повертає список коментарів для конкретного тикету
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID тикету
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
        - name: page
          in: query
          description: Номер сторінки
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Кількість записів на сторінку
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список коментарів
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentsListResponse'
        '401':
          description: Не авторизований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тикет не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Comments
      summary: Додавання коментаря
      description: Додає новий коментар до тикету
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID тикету
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
            example:
              content: "Оновив драйвери принтера. Тестування показало, що проблема вирішена."
              isInternal: false
      responses:
        '201':
          description: Коментар успішно додано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '400':
          description: Валідаційні помилки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизований
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тикет не знайдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439013"
        firstName:
          type: string
          example: "Іван"
        lastName:
          type: string
          example: "Петренко"
        email:
          type: string
          example: "ivan.petrenko@company.com"
        phone:
          type: string
          example: "+380501234567"
        role:
          type: string
          enum: [user, admin, support]
          example: "user"
        isActive:
          type: boolean
          example: true
        registrationStatus:
          type: string
          enum: [pending, approved, rejected]
          example: "approved"
...
    CreateCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          maxLength: 2000
          example: "Оновив драйвери принтера"
        isInternal:
          type: boolean
          default: false
          example: false
...
    Comment:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439018"
        content:
          type: string
          example: "Перевірив принтер, проблема в драйверах"
        author:
          $ref: '#/components/schemas/User'
        ticket:
          type: string
          example: "507f1f77bcf86cd799439017"
        isInternal:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 5
        totalItems:
          type: integer
          example: 47
        itemsPerPage:
          type: integer
          example: 10
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
    CommentResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Коментар успішно додано"
        data:
          $ref: '#/components/schemas/Comment'
    CommentsListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
            pagination:
              $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Помилка валідації"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Email обов'язковий"